Extract 'Shaped' interface
