+++ gon/geometry.py
@@ -105,54 +105,7 @@ class Linear(Geometry):
         """
 
 
-class ShapedCompound(Compound):
-    def __ge__(self, other: 'Geometry') -> bool:
-        """
-        Checks if the geometry is a superset of the other.
-        """
-        return (self is other
-                or (self.relate(other) in (Relation.EQUAL, Relation.COMPONENT,
-                                           Relation.ENCLOSED, Relation.WITHIN)
-                    if isinstance(other, Compound)
-                    else NotImplemented))
-
-    def __gt__(self, other: 'Geometry') -> bool:
-        """
-        Checks if the geometry is a strict superset of the other.
-        """
-        return (self is not other
-                and (self.relate(other) in (Relation.COMPONENT,
-                                            Relation.ENCLOSED, Relation.WITHIN)
-                     if isinstance(other, Compound)
-                     else NotImplemented))
-
-    def __le__(self, other: 'Geometry') -> bool:
-        """
-        Checks if the geometry is a subset of the other.
-        """
-        return (self is other
-                or ((self.relate(other) in (Relation.COVER, Relation.ENCLOSES,
-                                            Relation.COMPOSITE, Relation.EQUAL)
-                     if isinstance(other, ShapedCompound)
-                     # shaped cannot be subset of linear
-                     else False)
-                    if isinstance(other, Compound)
-                    else NotImplemented))
-
-    def __lt__(self, other: 'Geometry') -> bool:
-        """
-        Checks if the geometry is a strict subset of the other.
-        """
-        return (self is not other
-                and ((self.relate(other) in (Relation.COVER,
-                                             Relation.ENCLOSES,
-                                             Relation.COMPOSITE)
-                      if isinstance(other, ShapedCompound)
-                      # shaped cannot be strict subset of linear
-                      else False)
-                     if isinstance(other, Compound)
-                     else NotImplemented))
-
+class Shaped(Geometry):
     @property
     @abstractmethod
     def area(self) -> Coordinate:
